requirements.txt
colorama==0.3.9
Jinja2==2.9.6
MarkupSafe==1.0
pandoc==1.0.0b2
pandoc-attributes==0.1.7
pandocfilters==1.4.2
ply==3.10
PyYAML==3.12
six==1.10.0
